added example file
