+++ internal/app/storage/memory/mem_storage_test.go
@@ -2,7 +2,6 @@ package storage
 
 import (
 	"context"
-	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -63,24 +62,3 @@ func TestFindLink(t *testing.T) {
 }
 
 var testMemStorage = NewMemStorage()
-
-func ExampleMemStorage_PostLink() {
-	ctx := context.Background()
-
-	longLinkToSave := "https://practicum.yandex.ru/"
-
-	shortLink, _ := testMemStorage.PostLink(ctx, longLinkToSave, "http://localhost:8080", 0)
-	fmt.Println(shortLink)
-
-	// Output:
-	// http://localhost:8080/shortenLink1
-}
-
-func ExampleMemStorage_FindLink() {
-
-	longLink, _, _ := testMemStorage.FindLink("shortenLink1")
-	fmt.Println(longLink)
-
-	// Output:
-	// https://practicum.yandex.ru/
-}
